(No Logic) Move existing adjust-storage-dialog.component into new adjust-storage-dialog folder
