+++ apps/web/src/app/billing/individual/user-subscription.component.ts
@@ -15,7 +15,7 @@ import { DialogService, ToastService } from "@bitwarden/components";
 import {
   AdjustStorageDialogResult,
   openAdjustStorageDialog,
-} from "../shared/adjust-storage-dialog.component";
+} from "../shared/adjust-storage-dialog/adjust-storage-dialog.component";
 import {
   OffboardingSurveyDialogResultType,
   openOffboardingSurvey,
