+++ apps/web/src/app/billing/organizations/organization-subscription-cloud.component.ts
@@ -21,7 +21,7 @@ import { DialogService, ToastService } from "@bitwarden/components";
 import {
   AdjustStorageDialogResult,
   openAdjustStorageDialog,
-} from "../shared/adjust-storage-dialog.component";
+} from "../shared/adjust-storage-dialog/adjust-storage-dialog.component";
 import {
   OffboardingSurveyDialogResultType,
   openOffboardingSurvey,
